refactor(userspace-dp): #6386 extract host_inbound_policy sibling
Move the Junos host-inbound / to-zone junos-host policy helpers out of
poll_descriptor/mod.rs into a new pub(super) sibling
host_inbound_policy.rs: policy_packet_icmp, junos_host_policy_eval,
enum JunosHostLocalPolicy, emit_junos_host_deny, emit_host_inbound_deny,
and junos_host_local_policy. Each moved bare fn/enum becomes pub(super)
(they were module-private to mod.rs); nothing widens beyond pub(super).

Pure code motion. Bodies byte-identical to their prior location
(verified with git diff --color-moved). policy_packet_icmp keeps its
#[inline]; the two per-packet-capable evaluators junos_host_policy_eval
and junos_host_local_policy GAIN #[inline] (the only non-motion change
here — restores same-CGU inlining eligibility across the new module
boundary per the #6386 hot-path preservation contract); the two cold
event emitters stay un-hinted.

Cross-module edge made explicit: host_inbound_policy imports
enqueue_deny_reply from reject_reply, and the now-dead enqueue_deny_reply
import is pruned from mod.rs (its only caller moved). mod.rs re-imports
all six names via `use host_inbound_policy::{...}` — the loop uses four
directly and the still-resident flowless_local_delivery_verdict uses
junos_host_policy_eval + emit_junos_host_deny (both move out in the
flowless_verdict commit) — so every call site is textually unchanged.

Validation: cargo build (release) clean; moved bodies byte-identical
modulo pub(super) and the two #[inline] additions.
